requirements.txt
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
Markdown==3.4.1
pipreqs==0.4.11
django-cors-headers===3.10.0
psycopg2-binary==2.9.3
pandas==1.4.3
numpy==1.23.1
djangorestframework_simplejwt==5.2.0
PyJWT==2.4.0
gunicorn==20.1.0
django-redis==5.2.0